+++ plugins/colorizer/install.inc.php
@@ -1,8 +1,6 @@
 <?php
-
-$error = '';
-if ($error != '') {
-  $REX['ADDON']['installmsg']['colorizer'] = $error;
+if(isset($ADDONSsic) && $ADDONSsic['status']['website_manager']) { // workaround for OOAdon::isAvailable() for plugins
+	$REX['ADDON']['installmsg']['colorizer'] = $I18N->msg('be_utilities_uninstall_website_manager'); 
 } else {
   $REX['ADDON']['install']['colorizer'] = true;
 }