+++ plugins/colorizer/settings.inc.php
@@ -1,5 +1,5 @@
 <?php
 // --- DYN
-$REX['ADDON']['colorizer']['labelcolor'] = "#47a1ce";
-$REX['ADDON']['colorizer']['colorize_favicon'] = 0;
+$REX['ADDON']['colorizer']['labelcolor'] = "#b3b3b3";
+$REX['ADDON']['colorizer']['colorize_favicon'] = 1;
 // --- /DYN